move: 공유 모달 폴더 위치 변경
